xhs-content-extractor-for-mac/scripts/requirements.txt
requests>=2.32.0
selenium>=4.27.0
Pillow>=10.0.0
pyobjc-core>=11.1,<12
pyobjc-framework-Cocoa>=11.1,<12
pyobjc-framework-Quartz>=11.1,<12
pyobjc-framework-Vision>=11.1,<12
